app/requirements.txt
Pillow==10.4.0
ultralytics
onnxscript
numpy==1.26.4
jinja2==3.1.4
python-multipart==0.0.9
pycuda



